Implement a proper CFA abstraction
Create a new CFA struct that so far only handles normal 2x2 bayer
but that is based on a precalculated 48x48 matrix of colors so that
it works for 2x2, 6x6 (xtrans) and even 16x16 (even weirder stuff).

Light benchmarking says this is performance neutral against the
previous implementation with the added benefit that this will not
degrade performance to hand the non 2x2 cases.