demosaic: fractional scaling and fixup size calcs
Generalize the scaling code into fractional scaling factors by
selecting for each output pixel the set of input pixels that
belong to it. Right now all the pixels are weighed equally but
that should probably be fixed so the pixels at the edges of the
box are weighed less.

As for the scaling calculations they are now correct to the pixel
in that a NxM bounding box will always generate an image with one
of the sides equal to N or M.