imageops: split float conversion and crops
Add a new gofloat operation that does the u16->f32 conversion and
applies the crops. This simplifies the demosaic step considerably
and will in the future make it easy to handle native f32 images.

There's a small performance penalty of around 50ms from doing this
which seems low enough to be worth it for the simplicity. Even when
doing a very small thumbnail it's only a ~15% increase.