Avoid copying in the non-raw path
Keep the original DynamicImage in the pipeline with no conversions and
then do a conversion in gofloat before going to f32. Saves one clone().

Another copying can be saved by going directly from DynamicImage to the
f32 buffer. When the image crate adds f32 support that will be trivial
to do so just wait for a "to_rgbaf32()" method to appear. That would
give us an RGBA buffer already in [0;1] f32 that we can then just zero
out the alpha to get the appropriate RGBE.