Speedup XYZ to Lab with a lookup table
Use a lookup table to speed up the expensive cube root part of the XYZ
to Lab conversion. This required a 2^20*f32=4MB sized table to have
enough precision to still roundtrip. It may be possible to cut down the
size of this table by applying some kind of cheap non-linear transform
to the keys to be able to apply more bits to the most needed parts of
the range.