Start making rotation roundtrip transform sizing
Because of precision problems the roundtripping was highly imperfect,
particularly for small images and small rotations. So instead change the
contract around the transforms so that:

Both transform_forward() and transform_reverse() can save data, probably
mostly in private fields. This allows the reverse to use data saved
during the first forward (e.g., the aspect ratio of the input buffer).

Once transform_reverse() has been called and said what it needs to
create that output, then the next call to transform_forward() needs to
return that same buffer size. Most of the time this can be implemented
mathematically but here we just have transform_reverse() save the size
of the input buffer it got and once that happens transform_forward()
just returns that with no other calculation.